refactor(openomni): trim zero-consumer exports from the root barrel
Re-verified production consumers (apps/*, other packages, script/) for
every root-barrel value export and removed the 17 with zero production
consumers: IngressEventProjector, IngressHandlers,
IngressSessionResolver, resolveTarget, targetKey, PolicyResolver,
CronJobRegistry, defineTool, ToolRuntimePolicyMiddleware,
ReadBackExecutor, VerifierRegistry, CompletionSourceOrigin,
projectCompletionOrigin, DispatchPolicyRegistrationError,
DispatchRegistry, deriveActorContext, registerBuiltInDispatchHandlers.
With the three driver exports already evicted, the barrel goes from 55
to 35 value exports. Test-only consumers now deep-import the owning
module or cluster index; src/evidence/index.ts had no remaining
consumer and is deleted.

Kept, with reasons: IngressAuthorityMiddleware, Tool, createDispatchTool,
resolveMeta, createDefaultDispatchPolicy, createDeviceDispatchHandlers,
createOutboundDispatchHandlers (removal orphans re-exports in
ingress/execution-runtime/dispatch cluster indexes owned by other
leaves and would grow the dead-export ratchet) and VerifierConformance
(the entry re-export shields the #493-owned dormant replay vocabulary).
Dead-export ratchet unchanged at 24 known issues.
