--- packages/openomni/src/evidence/index.ts
@@ -1,3 +1,0 @@
-export { ReadBackExecutor } from "./read-back-executor";
-export { VerifierRegistry } from "./verifier-registry";
-export * as VerifierConformance from "./verifier-conformance";
