+++ packages/openomni/test/evidence/read-back-executor-record.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 import { Storage, WorkItemStore } from "@openomni/session";
-import { ReadBackExecutor } from "../../src/index";
+import { ReadBackExecutor } from "../../src/evidence/read-back-executor";
 import { cleanupReadBackFixtures, LOCAL_READ_BACK, startFixtureServer } from "./read-back-fixture";
 
 beforeEach(() => {
