+++ packages/openomni/test/evidence/read-back-fixture.ts
@@ -1,7 +1,7 @@
 import { createServer, type Server, type ServerResponse } from "node:http";
 import { expect } from "bun:test";
 import { Storage } from "@openomni/session";
-import type { ReadBackExecutor } from "../../src/index";
+import type { ReadBackExecutor } from "../../src/evidence/read-back-executor";
 
 const servers: Server[] = [];
 const hangingResponses: ServerResponse[] = [];
